move audio file tests to specific test module
